requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.56.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==1.6.0
packaging==24.2
pandas==2.0.3
pillow==11.1.0
praw==7.7.1
prawcore==2.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
